写了lineMap的实现类，2分法。 暂停，开始补AST的基础内容。

+++ internal/ast_tree2/line_map.go
@@ -9,7 +9,7 @@ type LineMap interface {
 	/**
 	 * Find the start position of a line.
 	 */
-	GetStartPosition(line int64) int
+	GetStartPosition(line int) int
 	/**
 	 * Find the position corresponding to a (line,column).
 	 *
@@ -21,7 +21,7 @@ type LineMap interface {
 	 *           if {@code line < 1}
 	 *           if {@code line > no. of lines}
 	 */
-	GetPosition(line int64, column int64) int64
+	GetPosition(line int, column int) int
 
 	/**
 	 * Find the line containing a position; a line termination
@@ -30,7 +30,7 @@ type LineMap interface {
 	 * @param   pos  character offset of the position
 	 * @return the line number of pos (first line is 1)
 	 */
-	GetLineNumber(pos int64) int64
+	GetLineNumber(pos int) int
 
 	/**
 	 * Find the column for a character position.
@@ -40,5 +40,5 @@ type LineMap interface {
 	 * @param  pos   character offset of the position
 	 * @return       the tab-expanded column number of pos (first column is 1)
 	 */
-	GetColumnNumber(pos int64) int64
+	GetColumnNumber(pos int) int
 }
